[TEST] Lotto rank 구현

+++ src/main/java/lotto/domain/LottoStatistics.java
@@ -16,7 +16,7 @@ public final class LottoStatistics {
         float totalPrizeAmount = rankings.entrySet().stream()
                 .mapToInt(iEntrySet ->
                     iEntrySet.getKey().prizeAmount(
-                            iEntrySet.getValue().intValue()
+                            iEntrySet.getValue()
                     ).intValue()
                 ).sum();
         float totalPurchaseAmount = Lotto.PRICE.multiply(totalSize()).intValue();
